requirements.txt
Django==2.2
djangorestframework==3.12.1
factory_boy==2.12.0